+++ pkg/ttnpb/identifiers.pb.paths.fm.go
@@ -95,3 +95,18 @@ var EntityIdentifiersFieldPathsNested = []string{
 var EntityIdentifiersFieldPathsTopLevel = []string{
 	"ids",
 }
+var EndDeviceVersionIdentifiersFieldPathsNested = []string{
+	"band_id",
+	"brand_id",
+	"firmware_version",
+	"hardware_version",
+	"model_id",
+}
+
+var EndDeviceVersionIdentifiersFieldPathsTopLevel = []string{
+	"band_id",
+	"brand_id",
+	"firmware_version",
+	"hardware_version",
+	"model_id",
+}