+++ CHANGELOG.md
@@ -13,6 +13,7 @@ For details about compatibility between different releases, see the **Commitment
 
 - Searching Packet Broker networks with `--tenant-id-contains` and `--name-contains` flags.
 - Listing all listed Packet Broker networks with `ttn-lw-cli packetbroker networks list`.
+- Include end device version identifiers in upstream messages (see `uplink_message.version_ids.brand_id`, `uplink_message.version_ids.model_id`, `uplink_message.version_ids.firmware_version`, `uplink_message.version_ids.hardware_version` and `uplink_message.version_ids.band_id` fields).
 
 ### Changed